fix: address PR #431 review feedback — 5 items resolved
1. build: add missing createdAt int64 param to panicIndex.IndexEngram (worker_test.go) and 5 test call sites (fts_test.go)
2. plasticity: revert default ScoringFusion from RRF back to ACT-R empty string
3. activation: merge RRF threshold default logic — only apply 0.001 when user has not explicitly set a threshold
4. search: gate bleve backend behind //go:build bleve tag, add search/factory package for compile-time dependency isolation
5. polish: clarify native DeleteText no-op comment, note HNSW metrics not applicable to bleve, move AGENTS.md to docs/
