+++ internal/search/bleve/backend_noop_test.go
@@ -1,4 +1,4 @@
-//go:build !vectors
+//go:build bleve && !vectors
 
 package bleve_test
