+++ internal/search/bleve/backend_test.go
@@ -1,3 +1,5 @@
+//go:build bleve
+
 package bleve_test
 
 import (
