+++ internal/search/native/backend.go
@@ -39,10 +39,12 @@ func (b *Backend) IndexEngram(ws [8]byte, id [16]byte, concept, createdBy, conte
 	return b.FTS.IndexEngram(ws, id, concept, createdBy, content, tags, 0)
 }
 
-// DeleteText removes the text index entries for id when enough source text is available.
+// DeleteText is intentionally a no-op for the native backend: the engine
+// calls fts.Index.DeleteEngram directly, where the original tokenized fields
+// are available to remove exact posting keys. Future contributors should not
+// mistake this lack of implementation for a bug — deletion happens at the
+// engine layer, not through this adapter.
 func (b *Backend) DeleteText(_ context.Context, ws [8]byte, id [16]byte) error {
-	// Native FTS deletion needs the original tokenized fields to remove exact posting keys.
-	// Existing engine deletion still calls fts.Index.DeleteEngram with that context.
 	return nil
 }
