+++ internal/transport/rest/search_backend_vectors_test.go
@@ -1,4 +1,4 @@
-//go:build vectors
+//go:build vectors && bleve
 
 package rest
